No need for a middle man script

+++ deploy.php
@@ -2,21 +2,36 @@
 <?php
 $argv = $_SERVER['argv'];
 
-$commands = array(
-    'cd ' . $argv[0],
-    'git reset --hard',
-    'git checkout ' . $argv[1],
-    'git fetch origin',  
-    'git rebase origin/'. $argv[1] . ' ' . $argv[1],
-    'git status',
-);
+if (file_exists($argv[0])) {
 
-$output = '';
-foreach($commands AS $command) {
-    $tmp = shell_exec($command . " 2>&1");
-    $output .= "{$command}\n";
-    $output .= htmlentities(trim($tmp)) . "\n";
+	//Get first line
+	$request = fopen($argv[0]);
+	$arguments = fgets($request);
+	fclose($request);
+
+	//Split string
+	$params = explode("|", $arguments);
+
+	//Run commands on repo
+	$commands = array(
+	    'cd ' . $params[0],
+	    'git reset --hard',
+	    'git checkout ' . $params[1],
+	    'git fetch origin',  
+	    'git rebase origin/'. $params[1] . ' ' . $params[1],
+	    'git status',
+	);
+
+	$output = '';
+	foreach($commands AS $command) {
+	    $tmp = shell_exec($command . " 2>&1"); //> /dev/null 2>&1 &
+	    $output .= "{$command}\n";
+	    $output .= htmlentities(trim($tmp)) . "\n";
+	}
+
+	echo $output;
+
+	//TODO: Delete request file
 }
 
-echo $output;
 ?>
